server/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-auth==2.23.0
requests==2.31.0
flask-cors==4.0.0
oauthlib==3.2.2
authlib==1.2.1 